fix: change gateway matcher to exclude url fragments & queries
Currently js.ipfs.io breaks when one tries to reload Peerpad because it stores information in the URL fragment (for example) (#blah) section of the URL, which this gateway tries to resolve as part of the IPFS path.

This makes the behavior consistent with the other gateways which ignore query strings and fragments.